Creating InitialScreen style file
