+++ src/screens/InitialScreen.js
@@ -1,57 +1,13 @@
 import React from 'react';
 import { Actions } from 'react-native-router-flux';
-import { StyleSheet,
+import {
   Text,
   View,
   ImageBackground,
   TouchableOpacity } from 'react-native';
+import styles from '../Styles/InitialScreenStyles';
 
 const initialBackgroundImage = require('../images/initial_background.jpg');
-
-const styles = StyleSheet.create({
-  initialBackgroundImage: {
-    flex: 1,
-    width: null,
-  },
-
-  initialScreen: {
-    flex: 1,
-    backgroundColor: 'transparent',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 1,
-  },
-
-  loginButton: {
-    width: 320,
-    height: 65,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#FF9500',
-    borderRadius: 10,
-    marginBottom: 60,
-  },
-
-  buttonText: {
-    fontSize: 20,
-    color: 'white',
-  },
-
-  registerButton: {
-    width: 320,
-    height: 65,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#FF9500',
-    borderRadius: 10,
-  },
-
-  titleText: {
-    fontSize: 60,
-    fontWeight: 'bold',
-    color: 'white',
-    marginTop: 140,
-  },
-});
 
 const InitialScreen = () => ({
   render() {
